refactor: move copy call link functionality to webrtcStore & set up a seperate smaller call link button when in the call
- Extract copyCallLink function from CallWindow component to webrtcStore for reusability
- Add copy call link button to MainCallControls when user is in call
- Hide large copy link button in CallWindow when user has joined call
- Add clipboard feedback state (copied/not copied) to both components
- Import generateCallInviteUrl utility in webrtcStore
- Remove unused imports from CallWindow component
