+++ app/src/components/call/window/CallWindow.vue
@@ -52,7 +52,7 @@
         <JoinCallControls v-if="!inCall" />
         <MainCallControls v-if="inCall" />
 
-        <j-button @click="copyCallLink" size="lg">
+        <j-button v-if="!inCall" @click="handleCopyCallLink" size="lg">
           <j-icon :name="hasCopiedLink ? 'clipboard-check' : 'link-45deg'" :style="{ '--j-icon-size': hasCopiedLink ? '1.5em' : '1.9em', margin: hasCopiedLink ? '0 -5px 0 0' : '0 -5px -3px 0' }" />
           Copy Call Invite Link
         </j-button>
@@ -76,11 +76,9 @@
 
 <script setup lang="ts">
 import AvatarGroup from '@/components/avatar-group/AvatarGroup.vue';
-import { useAppStore, useCommunityServiceStore, useUiStore, useWebrtcStore } from '@/stores';
-import { generateCallInviteUrl } from '@/utils/callInviteUrl';
+import { useUiStore, useWebrtcStore } from '@/stores';
 import { storeToRefs } from 'pinia';
 import { ref } from 'vue';
-import { useRoute } from 'vue-router';
 import { useCallResize } from '../composables/useCallResize';
 import JoinCallControls from '../controls/JoinCallControls.vue';
 import MainCallControls from '../controls/MainCallControls.vue';
@@ -95,14 +93,11 @@ defineProps<{
   };
 }>();
 
-const appStore = useAppStore();
 const uiStore = useUiStore();
 const webrtcStore = useWebrtcStore();
-const communityServiceStore = useCommunityServiceStore();
-const route = useRoute();
 
 const { callWindowWidth, callWindowOpen, isMobile } = storeToRefs(uiStore);
-const { agentsInCall, inCall, callRoute } = storeToRefs(webrtcStore);
+const { agentsInCall, inCall } = storeToRefs(webrtcStore);
 
 const rightSection = ref<HTMLElement | null>(null);
 const callWindow = ref<HTMLElement | null>(null);
@@ -114,38 +109,13 @@ function closeCallWindow() {
   uiStore.setCallWindowOpen(false);
 }
 
-async function copyCallLink() {
-  try {
-    // Get route params from current route instead of callRoute (which is only set when joined)
-    const communityId = route.params.communityId as string;
-    const channelId = route.params.channelId as string;
-    
-    if (!communityId || !channelId) {
-      appStore.showDangerToast({ message: 'Cannot generate call link - missing route information' });
-      return;
-    }
-
-    // Get the neighbourhood URL for this community
-    const communityService = communityServiceStore.getCommunityService(communityId);
-    const neighbourhoodUrl = communityService?.perspective.sharedUrl;
-
-    if (!neighbourhoodUrl) {
-      appStore.showDangerToast({ message: 'Cannot generate call link - community not found' });
-      return;
-    }
-
-    const callUrl = generateCallInviteUrl(neighbourhoodUrl, channelId);
-    await navigator.clipboard.writeText(callUrl);
-    
+async function handleCopyCallLink() {
+  const success = await webrtcStore.copyCallLink();
+  if (success) {
     hasCopiedLink.value = true;
-    appStore.showSuccessToast({ message: 'Call invite link copied to clipboard!' });
-    
     setTimeout(() => {
       hasCopiedLink.value = false;
     }, 3000);
-  } catch (error) {
-    console.error('Failed to copy call link:', error);
-    appStore.showDangerToast({ message: 'Failed to copy call link. Please try again.' });
   }
 }
 </script>
