+++ app/src/stores/webrtcStore.ts
@@ -4,6 +4,7 @@ import pigWav from '@/assets/audio/pig.wav';
 import popWav from '@/assets/audio/pop.wav';
 import { HEARTBEAT_INTERVAL } from '@/composables/useSignallingService';
 import { getCachedAgentProfile } from '@/utils/userProfileCache';
+import { generateCallInviteUrl } from '@/utils/callInviteUrl';
 import { PerspectiveExpression } from '@coasys/ad4m';
 import { AgentState, AgentStatus, CallHealth, Profile, RouteParams } from '@coasys/flux-types';
 import { Howl } from 'howler';
@@ -731,6 +732,36 @@ export const useWebrtcStore = defineStore(
       { immediate: true },
     );
 
+    async function copyCallLink(): Promise<boolean> {
+      try {
+        const communityId = callRoute.value.communityId || route.params.communityId as string;
+        const channelId = callRoute.value.channelId || route.params.channelId as string;
+        
+        if (!communityId || !channelId) {
+          appStore.showDangerToast({ message: 'Cannot generate call link - missing route information' });
+          return false;
+        }
+
+        const communityService = getCommunityService(communityId);
+        const neighbourhoodUrl = communityService?.perspective.sharedUrl;
+
+        if (!neighbourhoodUrl) {
+          appStore.showDangerToast({ message: 'Cannot generate call link - community not found' });
+          return false;
+        }
+
+        const callUrl = generateCallInviteUrl(neighbourhoodUrl, channelId);
+        await navigator.clipboard.writeText(callUrl);
+        
+        appStore.showSuccessToast({ message: 'Call invite link copied to clipboard!' });
+        return true;
+      } catch (error) {
+        console.error('Failed to copy call link:', error);
+        appStore.showDangerToast({ message: 'Failed to copy call link. Please try again.' });
+        return false;
+      }
+    }
+
     return {
       inCall,
       callRoute,
@@ -755,6 +786,7 @@ export const useWebrtcStore = defineStore(
       signalAgent,
       signalAgentsInCall,
       displayEmoji,
+      copyCallLink,
     };
   },
   { persist: false },
